Add solution for day 9 part 2
In addition, slightly rewrite part 1, because of rewriting shared module
code to return coordinates of neighbors instead of their values.

Still missing something important in recursive walk-through, because we
get doubles in the resulting visited list – which should be avoided –
and this is solved by additional `Enum.uniq/1` call before return.